Ent 12681 [#16804]Añadir decimales al modulo module_freepercentdisk

See merge request artica/pandorafms!6882

+++ pandora_agents/win32/windows/pandora_wmi.h
@@ -44,7 +44,7 @@ namespace Pandora_Wmi {
 	int           isProcessRunning      (string process_name);
 	int           isServiceRunning      (string service_name);
 	unsigned long getDiskFreeSpace      (string disk_id);
-	unsigned long getDiskFreeSpacePercent      (string disk_id);
+	double        getDiskFreeSpacePercent      (string disk_id);
 	int           getCpuUsagePercentage (int cpu_id);
 	long          getFreememory         ();
 	long          getFreememoryPercent  ();
